feat: extract markdown storage into standalone crate
