+++ Cargo.toml
@@ -29,6 +29,7 @@ members = [
     "vtcode-commons",
     "vtcode-config",
     "vtcode-llm",
+    "vtcode-markdown-store",
     "vtcode-tools",
 ]
