+++ docs/component_extraction_todo.md
@@ -29,6 +29,13 @@ This list tracks actionable tasks spawned from the component extraction plan as 
 - [x] Publish a runnable example that demonstrates installing custom defaults providers when loading configuration outside `.vtcode` directories.
 
 
+## `vtcode-markdown-store`
+- [x] Scaffold the standalone crate exporting markdown storage, key-value, project, and cache helpers under optional feature flags.
+- [x] Add constructors that accept custom project roots so `.vtcode` directory assumptions are no longer hardcoded.
+- [ ] Document crate usage patterns and feature gates for downstream adopters before publishing an initial release.
+- [ ] Investigate file locking or atomic write strategies to support concurrent agents sharing the same storage root.
+
+
 ## Cross-Cutting
 - [x] Capture external integration prerequisites (environment variables, binary dependencies) and align them with optional feature groups.
 - [x] Define shared base traits (`vtcode-commons`) for filesystem paths, telemetry, and error handling used by multiple crates.
