+++ vtcode-core/Cargo.toml
@@ -112,6 +112,7 @@ portable-pty = "0.9.0"
 ansi-to-tui = "7.0.0"
 vtcode-commons = { path = "../vtcode-commons" }
 vtcode-config = { path = "../vtcode-config" }
+vtcode-markdown-store = { path = "../vtcode-markdown-store" }
 
 # Token counting for attention budget management
 tokenizers = { version = "0.15", features = ["http"] }
